+++ src/engines/cse/environment.ts
@@ -3,7 +3,6 @@ import { MissingRequiredPositionalError, TooManyPositionalArgumentsError } from 
 import { Closure } from "./closure";
 import { Context } from "./context";
 import { handleRuntimeError } from "./error";
-import { Heap } from "./heap";
 import { Value } from "./stash";
 
 export interface Frame {
@@ -38,9 +37,6 @@ export interface Environment {
    */
   head: Frame;
 
-  heap: Heap;
-  thisContext?: Value;
-
   /**
    * The closure associated with this environment, if this environment was created as part of a function call.
    */
@@ -65,14 +61,11 @@ export const createEnvironment = (
         : "lambda",
     tail: closure.environment,
     head: {},
-    heap: new Heap(),
     id: uniqueId(context),
     callExpression: callExpression,
     closure: closure,
   };
-
-  // console.info('closure.node.params:', closure.node.params);
-  // console.info('Number of params:', closure.node.params.length);
+  
   const isVariadic = closure.node.parameters.some(param => param.isStarred);
   let consumed = false;
   closure.node.parameters.forEach((paramToken, index) => {
@@ -133,8 +126,7 @@ export const createSimpleEnvironment = (
     name,
     tail,
     head: {},
-    heap: new Heap(),
-    // TODO: callExpression and thisContext are optional and can be provided as needed.
+    // TODO: callExpression is optional and can be provided as needed.
   };
 };
 
@@ -154,23 +146,9 @@ export const createBlockEnvironment = (
     name,
     tail: currentEnvironment(context),
     head: {},
-    heap: new Heap(),
     id: uniqueId(context),
   };
 };
-
-// export const handleArrayCreation = (
-//   context: Context,
-//   array: Value[],
-//   envOverride?: Environment
-// ): void => {
-//   const environment = envOverride ?? currentEnvironment(context)
-//   Object.defineProperties(array, {
-//     id: { value: uniqueId(context) },
-//     environment: { value: environment, writable: true }
-//   })
-//   environment.heap.add(array)
-// }
 
 export const currentEnvironment = (context: Context): Environment => {
   return context.runtime.environments[0];
@@ -190,5 +168,4 @@ export const popEnvironment = (context: Context) => context.runtime.environments
 
 export const pushEnvironment = (context: Context, environment: Environment) => {
   context.runtime.environments.unshift(environment);
-  context.runtime.environmentTree.insert(environment);
 };
