+++ src/engines/cse/types.ts
@@ -1,5 +1,5 @@
 import { ExprNS, StmtNS } from "../../ast-types";
-import { TokenType } from "../../tokens";
+import { TokenType } from "../../tokenizer";
 import { Environment } from "./environment";
 import { Value } from "./stash";
