+++ src/engines/wasm/builderGenerator.ts
@@ -1,55 +1,55 @@
 import {
-  f64,
-  global,
-  i32,
-  i64,
-  local,
-  mut,
-  wasm,
-  WasmCall,
-  WasmData,
-  WasmExport,
-  type WasmInstruction,
-  type WasmNumeric,
-  type WasmRaw,
+    f64,
+    global,
+    i32,
+    i64,
+    local,
+    mut,
+    wasm,
+    WasmCall,
+    WasmData,
+    WasmExport,
+    type WasmInstruction,
+    type WasmNumeric,
+    type WasmRaw,
 } from "@sourceacademy/wasm-util";
 import { WasmExports } from ".";
 import { ExprNS, StmtNS } from "../../ast-types";
-import { TokenType } from "../../tokens";
+import { TokenType } from "../../tokenizer";
 import {
-  ALLOC_ENV_FX,
-  APPLY_FX_NAME,
-  applyFuncFactory,
-  ARITHMETIC_OP_FX,
-  ARITHMETIC_OP_TAG,
-  BOOL_NOT_FX,
-  BOOLISE_FX,
-  COMPARISON_OP_FX,
-  COMPARISON_OP_TAG,
-  CURR_ENV,
-  ENV_HEAD_SIZE,
-  GET_LEX_ADDR_FX,
-  GET_LIST_ELEMENT_FX,
-  HEAP_PTR,
-  importedLogs,
-  LOG_FX,
-  MAKE_BOOL_FX,
-  MAKE_CLOSURE_FX,
-  MAKE_COMPLEX_FX,
-  MAKE_FLOAT_FX,
-  MAKE_INT_FX,
-  MAKE_LIST_FX,
-  MAKE_NONE_FX,
-  MAKE_PAIR_FX,
-  MAKE_STRING_FX,
-  MALLOC_FX,
-  nativeFunctions,
-  NEG_FX,
-  PRE_APPLY_FX,
-  RETURN_ENV_NAME,
-  SET_CONTIGUOUS_BLOCK_FX,
-  SET_LEX_ADDR_FX,
-  SET_LIST_ELEMENT_FX,
+    ALLOC_ENV_FX,
+    APPLY_FX_NAME,
+    applyFuncFactory,
+    ARITHMETIC_OP_FX,
+    ARITHMETIC_OP_TAG,
+    BOOL_NOT_FX,
+    BOOLISE_FX,
+    COMPARISON_OP_FX,
+    COMPARISON_OP_TAG,
+    CURR_ENV,
+    ENV_HEAD_SIZE,
+    GET_LEX_ADDR_FX,
+    GET_LIST_ELEMENT_FX,
+    HEAP_PTR,
+    importedLogs,
+    LOG_FX,
+    MAKE_BOOL_FX,
+    MAKE_CLOSURE_FX,
+    MAKE_COMPLEX_FX,
+    MAKE_FLOAT_FX,
+    MAKE_INT_FX,
+    MAKE_LIST_FX,
+    MAKE_NONE_FX,
+    MAKE_PAIR_FX,
+    MAKE_STRING_FX,
+    MALLOC_FX,
+    nativeFunctions,
+    NEG_FX,
+    PRE_APPLY_FX,
+    RETURN_ENV_NAME,
+    SET_CONTIGUOUS_BLOCK_FX,
+    SET_LEX_ADDR_FX,
+    SET_LIST_ELEMENT_FX,
 } from "./constants";
 import { LibFuncType } from "./library";
