+++ src/engines/wasm/metacircularGenerator.ts
@@ -1,6 +1,6 @@
 import { PARSE_TREE_STRINGS, WasmExports } from ".";
 import { ExprNS, StmtNS } from "../../ast-types";
-import { TokenType } from "../../tokens";
+import { TokenType } from "../../tokenizer";
 
 interface BuilderVisitor<S, E> extends StmtNS.Visitor<S>, ExprNS.Visitor<E> {
   visit(stmt: StmtNS.Stmt): S;
