+++ src/errors/errors.ts
@@ -1,12 +1,8 @@
 import { ExprNS, StmtNS } from "../ast-types";
-import { isStatementSequence } from "../engines/cse/closure";
 import { Context } from "../engines/cse/context";
-import { ControlItem } from "../engines/cse/control";
 import { operatorTranslator } from "../engines/cse/types";
-import { isNode } from "../engines/cse/utils";
 import { Token } from "../tokenizer";
-import { TokenType } from "../tokens";
-
+import { TokenType } from "../tokenizer/tokenizer";
 export enum ErrorType {
   IMPORT = "Import",
   RUNTIME = "Runtime",
@@ -380,21 +376,16 @@ export class ZeroDivisionError extends RuntimeSourceError {
 }
 
 export class StepLimitExceededError extends RuntimeSourceError {
-  constructor(source: string, node: ControlItem) {
-    const srcNode = isNode(node) ? node : node.srcNode;
-    const locatable = isStatementSequence(srcNode)
-        ? srcNode.body[0]
-        : srcNode;
-
-    super(locatable);
+  constructor(source: string, node: ExprNS.Expr | StmtNS.Stmt) {
+    super(node);
     this.type = ErrorType.RUNTIME;
-    const index = locatable.startToken.indexInSource;
+    const index = node.startToken.indexInSource;
 
     const { lineIndex, fullLine } = getFullLine(source, index);
 
     const errorPos =
-      "operator" in locatable && locatable.operator instanceof Token
-        ? locatable.operator.indexInSource - locatable.startToken.indexInSource
+      "operator" in node && node.operator instanceof Token
+        ? node.operator.indexInSource - node.startToken.indexInSource
         : 0;
 
     const indicator = createErrorIndicator(fullLine, errorPos); // no target symbol
