--- src/tokens.ts
@@ -1,98 +1,0 @@
-export enum TokenType {
-  ENDMARKER = 0,
-  NAME = 1,
-  NUMBER = 2,
-  BIGINT = 3,
-  STRING = 4,
-  NEWLINE = 5,
-  INDENT = 6,
-  DEDENT = 7,
-  LPAR = 8,
-  RPAR = 9,
-  COLON = 10,
-  DOUBLECOLON = 11,
-  COMMA = 12,
-  PLUS = 13,
-  MINUS = 14,
-  BANG = 15,
-  STAR = 16,
-  SLASH = 17,
-  VBAR = 18,
-  AMPER = 19,
-  LESS = 20,
-  GREATER = 21,
-  EQUAL = 22,
-  PERCENT = 23,
-  DOUBLEEQUAL = 24,
-  NOTEQUAL = 25,
-  LESSEQUAL = 26,
-  GREATEREQUAL = 27,
-  DOUBLESTAR = 28,
-  COMPLEX = 29,
-  AND = 30,
-  OR = 31,
-  FOR = 32,
-  WHILE = 33,
-  NONE = 34,
-  TRUE = 35,
-  FALSE = 36,
-  IS = 37,
-  NOT = 38,
-  ISNOT = 39,
-  PASS = 40,
-  DEF = 41,
-  LAMBDA = 42,
-  FROM = 43,
-  DOUBLESLASH = 44,
-  BREAK = 45,
-  CONTINUE = 46,
-  RETURN = 47,
-  ASSERT = 48,
-  IMPORT = 49,
-  GLOBAL = 50,
-  NONLOCAL = 51,
-  IF = 52,
-  ELSE = 53,
-  ELIF = 54,
-  IN = 55,
-  NOTIN = 56,
-  RSQB = 57,
-  LSQB = 58,
-  ELLIPSIS = 59,
-  SEMI = 60,
-  DOT = 61,
-  LBRACE = 62,
-  RBRACE = 63,
-  TILDE = 64,
-  CIRCUMFLEX = 65,
-  LEFTSHIFT = 66,
-  RIGHTSHIFT = 67,
-  PLUSEQUAL = 68,
-  MINEQUAL = 69,
-  STAREQUAL = 70,
-  SLASHEQUAL = 71,
-  PERCENTEQUAL = 72,
-  AMPEREQUAL = 73,
-  VBAREQUAL = 74,
-  CIRCUMFLEXEQUAL = 75,
-  LEFTSHIFTEQUAL = 76,
-  RIGHTSHIFTEQUAL = 77,
-  DOUBLESTAREQUAL = 78,
-  DOUBLESLASHEQUAL = 79,
-  AT = 80,
-  ATEQUAL = 81,
-  RARROW = 82,
-  COLONEQUAL = 83,
-  OP = 84,
-  AWAIT = 85,
-  ASYNC = 86,
-  TYPE_IGNORE = 87,
-  TYPE_COMMENT = 88,
-  YIELD = 89,
-  WITH = 90,
-  DEL = 91,
-  TRY = 92,
-  EXCEPT = 93,
-  FINALLY = 94,
-  RAISE = 95,
-}
